feat: Add display field, remove niri/labwc
Remove labwc/niri because the former isn't packaged and per consensus on the operations chat for niri.

Given that the filter for whether or not to show a given group is getting a little bit unwieldy add a new field "display" which when true will show a given group as a DE/WM option.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>